Abstract out repo input text field popover
